Add guide for installing Git on Linux/Windows/Mac
Adds a guide for students on how to install git on linux
window and mac.

Because the installation process is somewhat complicated on
Windows, this walks through each step, with a recommendation of the
settings to apply along the way.

Adds instructions for installing Git on Linux, and uses the
manual's instructions for mac. We need someone to test this for us.